feature user role permission

+++ app/Http/Controllers/Backend/SystemController.php
@@ -4,9 +4,11 @@ namespace App\Http\Controllers\Backend;
 
 use App\Consts\PerPage;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\SystemRequest;
 use App\Http\Resources\SystemResource;
 use App\Models\Filters\SystemFilter;
 use App\Models\System;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 class SystemController extends Controller
@@ -26,17 +28,19 @@ class SystemController extends Controller
 
     public function store(SystemRequest $request)
     {
+        $this->authorize('create', User::class);
         try {
             $data = $request->validated();
             System::create($data);
             return $this->sendSuccessResponse(null, __('Created successfully'));
-        } catch (Throwable $e) {
+        } catch (\Throwable $e) {
             return $this->sendErrorResponse(__('Something went wrong'), $e->getMessage());
         }
     }
 
     public function update($id, SystemRequest $request)
     {
+        $this->authorize('update', User::class);
         $data = $request->validated();
         try {
             $system = System::find($id);
@@ -45,28 +49,36 @@ class SystemController extends Controller
             }
             $system->update($data);
             return $this->sendSuccessResponse(null, __('Updated successfully'));
-        } catch (Throwable $e) {
+        } catch (\Throwable $e) {
             return $this->sendErrorResponse(__('Something went wrong'), $e->getMessage());
         }
     }
 
     public function show($id)
     {
+        $this->authorize('view', User::class);
         try {
             $system = System::find($id);
+            if(!$system) {
+                return $this->sendErrorResponse(__('Data not found'), 404);
+            }
             return $this->sendSuccessResponse(SystemResource::make($system));
-        } catch (Throwable $e) {
+        } catch (\Throwable $e) {
             return $this->sendErrorResponse(__('Something went wrong'), $e->getMessage());
         }
     }
 
     public function destroy($id)
     {
+        $this->authorize('delete', User::class);
         try {
             $system = System::find($id);
+            if(!$system) {
+                return $this->sendErrorResponse(__('Data not found'), 404);
+            }
             $system->delete();
             return $this->sendSuccessResponse(null, __('Deleted successfully'));
-        } catch (Throwable $e) {
+        } catch (\Throwable $e) {
             return $this->sendErrorResponse(__('Something went wrong'), $e->getMessage());
         }
     }
